tracer_eq_2d: overhaul - Combine conservative and non-conservative modes - Make TracerTerms label and index aware - Combine the TracerEquation2D interface - Allow systems of tracer equations solved as monolithic systems - Pass a suffix for each tracer term
