+++ thetis/tracer_eq_2d.py
@@ -12,6 +12,18 @@ The advection-diffusion equation of tracer :math:`T` in non-conservative form re
 where :math:`\nabla_h` denotes horizontal gradient, :math:`\textbf{u}` are the horizontal
 velocities, and
 :math:`\mu_h` denotes horizontal diffusivity.
+
+The advection-diffusion equation of depth-integrated tracer :math:`q=HT` in conservative form reads
+
+.. math::
+    \frac{\partial q}{\partial t}
+    + \nabla_h \cdot (\textbf{u} q)
+    = \nabla_h \cdot (\mu_h \nabla_h q)
+    :label: cons_tracer_eq_2d
+
+where :math:`\nabla_h` denotes horizontal gradient, :math:`\textbf{u}` are the horizontal
+velocities, and
+:math:`\mu_h` denotes horizontal diffusivity.
 """
 from .utility import *
 from .equation import Term, Equation
@@ -22,6 +34,10 @@ __all__ = [
     'HorizontalAdvectionTerm',
     'HorizontalDiffusionTerm',
     'SourceTerm',
+    'ConservativeTracerTerm',
+    'ConservativeHorizontalAdvectionTerm',
+    'ConservativeHorizontalDiffusionTerm',
+    'ConservativeSourceTerm',
 ]
 
 
@@ -30,15 +46,22 @@ class TracerTerm(Term):
     Generic tracer term that provides commonly used members and mapping for
     boundary functions.
     """
-    def __init__(self, function_space, depth, options, test_function=None):
+    def __init__(self, index, label, function_space, depth, options,
+                 test_function=None, trial_function=None):
         """
+        :arg index: index for this tracer component within the vector
+        :arg label: label for this tracer component
         :arg function_space: :class:`FunctionSpace` where the solution belongs
         :arg depth: :class:`DepthExpression` containing depth info
         :arg options: :class`ModelOptions2d` containing parameters
-        :kwarg test_function: custom :class:`TestFunction`.
+        :kwarg test_function: custom :class:`TestFunction`
+        :kwarg trial_function: custom :class:`TrialFunction`
         """
         super(TracerTerm, self).__init__(function_space,
-                                         test_function=test_function)
+                                         test_function=test_function,
+                                         trial_function=trial_function)
+        self.index = index
+        self.label = label
         self.depth = depth
         self.options = options
         self.cellsize = CellSize(self.mesh)
@@ -91,6 +114,12 @@ class TracerTerm(Term):
 
         return c_ext, uv_ext, elev_ext
 
+    def component(self, f):
+        """
+        Return the component of a function corresponding to this tracer.
+        """
+        return f if len(f.dof_dset.dim) == 1 else split(f)[self.index]
+
 
 class HorizontalAdvectionTerm(TracerTerm):
     r"""
@@ -120,14 +149,15 @@ class HorizontalAdvectionTerm(TracerTerm):
             return 0
         elev = fields_old['elev_2d']
         self.corr_factor = fields_old.get('tracer_advective_velocity_factor')
+        c = self.component(solution)
 
         uv = self.corr_factor * fields_old['uv_2d']
         # FIXME is this an option?
         lax_friedrichs_factor = fields_old.get('lax_friedrichs_tracer_scaling_factor')
 
         f = 0
-        f += -(Dx(uv[0] * self.test, 0) * solution
-               + Dx(uv[1] * self.test, 1) * solution) * self.dx
+        f += -(Dx(uv[0] * self.test, 0) * c
+               + Dx(uv[1] * self.test, 1) * c) * self.dx
 
         if self.horizontal_dg:
             # add interface term
@@ -135,19 +165,19 @@ class HorizontalAdvectionTerm(TracerTerm):
             un_av = (uv_av[0]*self.normal('-')[0]
                      + uv_av[1]*self.normal('-')[1])
             s = 0.5*(sign(un_av) + 1.0)
-            c_up = solution('-')*s + solution('+')*(1-s)
+            c_up = c('-')*s + c('+')*(1-s)
 
             f += c_up*(jump(self.test, uv[0] * self.normal[0])
                        + jump(self.test, uv[1] * self.normal[1])) * self.dS
             # Lax-Friedrichs stabilization
             if self.options.use_lax_friedrichs_tracer:
                 gamma = 0.5*abs(un_av)*lax_friedrichs_factor
-                f += gamma*dot(jump(self.test), jump(solution))*self.dS
+                f += gamma*dot(jump(self.test), jump(c))*self.dS
 
         for bnd_marker in self.boundary_markers:
             funcs = bnd_conditions.get(bnd_marker)
             ds_bnd = ds(int(bnd_marker), degree=self.quad_degree)
-            c_in = solution
+            c_in = c
             if funcs is not None:
                 c_ext, uv_ext, eta_ext = self.get_bnd_functions(c_in, uv, elev, bnd_marker, bnd_conditions)
                 uv_av = 0.5*(uv + uv_ext)
@@ -194,13 +224,14 @@ class HorizontalDiffusionTerm(TracerTerm):
     https://dial.uclouvain.be/pr/boreal/object/boreal:128254/
     """
     def residual(self, solution, solution_old, fields, fields_old, bnd_conditions):
-        if fields_old.get('diffusivity_h') is None:
+        diffusivity_h = fields_old.get(f'diffusivity_h-{self.label}')
+        if diffusivity_h is None:
             return 0
-        diffusivity_h = fields_old['diffusivity_h']
+        c = self.component(solution)
         diff_tensor = as_matrix([[diffusivity_h, 0, ],
                                  [0, diffusivity_h, ]])
         grad_test = grad(self.test)
-        diff_flux = dot(diff_tensor, grad(solution))
+        diff_flux = dot(diff_tensor, grad(c))
         sipg_factor = self.options.sipg_factor_tracer
 
         f = 0
@@ -219,16 +250,16 @@ class HorizontalDiffusionTerm(TracerTerm):
             ds_interior = self.dS
             f += sigma_max * inner(
                 jump(self.test, self.normal),
-                dot(avg(diff_tensor), jump(solution, self.normal)))*ds_interior
+                dot(avg(diff_tensor), jump(c, self.normal)))*ds_interior
             f += -inner(avg(dot(diff_tensor, grad(self.test))),
-                        jump(solution, self.normal))*ds_interior
+                        jump(c, self.normal))*ds_interior
             f += -inner(jump(self.test, self.normal),
-                        avg(dot(diff_tensor, grad(solution))))*ds_interior
+                        avg(dot(diff_tensor, grad(c))))*ds_interior
 
         for bnd_marker in self.boundary_markers:
             funcs = bnd_conditions.get(bnd_marker)
             ds_bnd = ds(int(bnd_marker), degree=self.quad_degree)
-            c_in = solution
+            c_in = c
             elev = fields_old['elev_2d']
             self.corr_factor = fields_old.get('tracer_advective_velocity_factor')
             uv = self.corr_factor * fields_old['uv_2d']
@@ -261,43 +292,210 @@ class SourceTerm(TracerTerm):
     """
     def residual(self, solution, solution_old, fields, fields_old, bnd_conditions):
         f = 0
-        source = fields_old.get('source')
+        source = fields_old.get(f'source-{self.label}')
         if source is not None:
             f += -inner(source, self.test)*self.dx
         return -f
 
 
+class ConservativeTracerTerm(TracerTerm):
+    """
+    Generic depth-integrated tracer term that provides commonly used members and mapping for
+    boundary functions.
+    """
+    def __init__(self, index, label, function_space, depth, options,
+                 test_function=None, trial_function=None):
+        """
+        :arg index: index for this tracer component within the vector
+        :arg label: label for this tracer component
+        :arg function_space: :class:`FunctionSpace` where the solution belongs
+        :arg depth: :class: `DepthExpression` containing depth info
+        :arg options: :class`ModelOptions2d` containing parameters
+        :kwarg test_function: custom :class:`TestFunction`.
+        :kwarg trial_function: custom :class:`TrialFunction`.
+        """
+        super(ConservativeTracerTerm, self).__init__(index, label, function_space, depth, options,
+                                                     test_function=test_function,
+                                                     trial_function=trial_function)
+
+    # TODO: at the moment this is the same as TracerTerm, but we probably want to overload its
+    # get_bnd_functions method
+
+
+class ConservativeHorizontalAdvectionTerm(ConservativeTracerTerm):
+    r"""
+    Advection of tracer term, :math:`\nabla \cdot \bar{\textbf{u}} \nabla q`
+
+    The weak form is
+
+    .. math::
+        \int_\Omega \boldsymbol{\psi} \nabla\cdot \bar{\textbf{u}} \nabla q  dx
+        = - \int_\Omega \left(\nabla_h \boldsymbol{\psi})\right) \cdot \bar{\textbf{u}} \cdot q dx
+        + \int_\Gamma \text{avg}(q\bar{\textbf{u}}\cdot\textbf{n}) \cdot \text{jump}(\boldsymbol{\psi}) dS
+
+    where the right hand side has been integrated by parts;
+    :math:`\textbf{n}` is the unit normal of
+    the element interfaces, and :math:`\text{jump}` and :math:`\text{avg}` denote the
+    jump and average operators across the interface.
+    """
+    def residual(self, solution, solution_old, fields, fields_old, bnd_conditions=None):
+        if fields_old.get('uv_2d') is None:
+            return 0
+        elev = fields_old['elev_2d']
+        self.corr_factor = fields_old.get('tracer_advective_velocity_factor')
+        c = self.component(solution)
+
+        uv = self.corr_factor * fields_old['uv_2d']
+        uv_p1 = fields_old.get('uv_p1')
+        uv_mag = fields_old.get('uv_mag')
+
+        lax_friedrichs_factor = fields_old.get('lax_friedrichs_tracer_scaling_factor')
+
+        f = 0
+        f += -(Dx(self.test, 0) * uv[0] * c
+               + Dx(self.test, 1) * uv[1] * c) * self.dx
+
+        if self.horizontal_dg:
+            # add interface term
+            uv_av = avg(uv)
+            un_av = (uv_av[0]*self.normal('-')[0]
+                     + uv_av[1]*self.normal('-')[1])
+            s = 0.5*(sign(un_av) + 1.0)
+            flux_up = c('-')*uv('-')*s + c('+')*uv('+')*(1-s)
+
+            f += (flux_up[0] * jump(self.test, self.normal[0])
+                  + flux_up[1] * jump(self.test, self.normal[1])) * self.dS
+            # Lax-Friedrichs stabilization
+            if self.options.use_lax_friedrichs_tracer:
+                if uv_p1 is not None:
+                    gamma = 0.5*abs((avg(uv_p1)[0]*self.normal('-')[0]
+                                     + avg(uv_p1)[1]*self.normal('-')[1]))*lax_friedrichs_factor
+                elif uv_mag is not None:
+                    gamma = 0.5*avg(uv_mag)*lax_friedrichs_factor
+                else:
+                    gamma = 0.5*abs(un_av)*lax_friedrichs_factor
+                f += gamma*dot(jump(self.test), jump(c))*self.dS
+        if bnd_conditions is not None:
+            for bnd_marker in self.boundary_markers:
+                funcs = bnd_conditions.get(bnd_marker)
+                ds_bnd = ds(int(bnd_marker), degree=self.quad_degree)
+                c_in = c
+                if funcs is not None:
+                    c_ext, uv_ext, eta_ext = self.get_bnd_functions(c_in, uv, elev, bnd_marker, bnd_conditions)
+                    uv_av = 0.5*(uv + uv_ext)
+                    un_av = self.normal[0]*uv_av[0] + self.normal[1]*uv_av[1]
+                    s = 0.5*(sign(un_av) + 1.0)
+                    flux_up = c_in*uv*s + c_ext*uv_ext*(1-s)
+                    f += (flux_up[0]*self.normal[0]
+                          + flux_up[1]*self.normal[1])*self.test*ds_bnd
+                else:
+                    f += c_in * (uv[0]*self.normal[0]
+                                 + uv[1]*self.normal[1])*self.test*ds_bnd
+
+        return -f
+
+
+class ConservativeHorizontalDiffusionTerm(ConservativeTracerTerm, HorizontalDiffusionTerm):
+    r"""
+    Horizontal diffusion term :math:`-\nabla_h \cdot (\mu_h \nabla_h q)`
+
+    Using the symmetric interior penalty method the weak form becomes
+
+    .. math::
+        -\int_\Omega \nabla_h \cdot (\mu_h \nabla_h q) \phi dx
+        =& \int_\Omega \mu_h (\nabla_h \phi) \cdot (\nabla_h q) dx \\
+        &- \int_{\mathcal{I}_h\cup\mathcal{I}_v} \text{jump}(\phi \textbf{n}_h)
+        \cdot \text{avg}(\mu_h \nabla_h q) dS
+        - \int_{\mathcal{I}_h\cup\mathcal{I}_v} \text{jump}(q \textbf{n}_h)
+        \cdot \text{avg}(\mu_h  \nabla \phi) dS \\
+        &+ \int_{\mathcal{I}_h\cup\mathcal{I}_v} \sigma \text{avg}(\mu_h) \text{jump}(q \textbf{n}_h) \cdot
+            \text{jump}(\phi \textbf{n}_h) dS \\
+        &- \int_\Gamma \mu_h (\nabla_h \phi) \cdot \textbf{n}_h ds
+
+    where :math:`\sigma` is a penalty parameter, see Hillewaert (2013).
+
+    Hillewaert, Koen (2013). Development of the discontinuous Galerkin method
+    for high-resolution, large scale CFD and acoustics in industrial
+    geometries. PhD Thesis. Université catholique de Louvain.
+    https://dial.uclouvain.be/pr/boreal/object/boreal:128254/
+    """
+    # TODO: at the moment the same as HorizontalDiffusionTerm
+    # do we need additional H-derivative term?
+    # would also become different if ConservativeTracerTerm gets different bc options
+
+
+class ConservativeSourceTerm(ConservativeTracerTerm):
+    r"""
+    Generic source term
+
+    The weak form reads
+
+    .. math::
+        F_s = \int_\Omega \sigma \phi dx
+
+    where :math:`\sigma` is a user defined scalar :class:`Function`.
+
+    """
+    def residual(self, solution, solution_old, fields, fields_old, bnd_conditions=None):
+        f = 0
+        source = fields_old.get(f'source-{self.label}')
+        if source is not None:
+            H = self.depth.get_total_depth(fields_old['elev_2d'])
+            f += -inner(H*source, self.test)*self.dx
+        return -f
+
+
 class TracerEquation2D(Equation):
     """
-    2D tracer advection-diffusion equation :eq:`tracer_eq` in conservative form
+    Several 2D tracer advection-diffusion equations :eq:`tracer_eq` in
+    conservative or non-conservative form, solved as a mixed system.
     """
-    def __init__(self, function_space, depth, options, velocity):
+    def __init__(self, system, function_space, depth, options, velocity):
         """
+        :arg system: comma separated tracer fields under consideration
         :arg function_space: :class:`FunctionSpace` where the solution belongs
         :arg depth: :class: `DepthExpression` containing depth info
         :arg options: :class`ModelOptions2d` containing parameters
         :arg velocity: velocity field associated with the shallow water model
         """
-        super(TracerEquation2D, self).__init__(function_space)
+        super().__init__(function_space)
+        self.depth = depth
+        self.options = options
+        self.velocity = velocity
+        test_functions = TestFunctions(function_space)
+        trial_functions = TrialFunctions(function_space)
+        for i, label in enumerate(system.split(',')):
+            args = (i, label, function_space[i], depth, options)
+            kwargs = {'trial_function': trial_functions[i]}
+            if options.use_supg_tracer:
+                # TODO: allow SUPG for only some fields
+                kwargs['test_function'] = self.apply_supg(test_functions[i])
+            else:
+                kwargs['test_function'] = test_functions[i]
+            if options.tracer[label].use_conservative_form:
+                self.add_conservative_terms(*args, **kwargs)
+            else:
+                self.add_nonconservative_terms(*args, **kwargs)
 
-        # Apply SUPG stabilisation
-        kwargs = {}
-        if options.use_supg_tracer:
-            unorm = options.horizontal_velocity_scale
-            if unorm.values()[0] > 0:
-                self.cellsize = anisotropic_cell_size(function_space.mesh())
-                tau = 0.5*self.cellsize/unorm
-                D = options.horizontal_diffusivity_scale
-                if D.values()[0] > 0:
-                    Pe = 0.5*unorm*self.cellsize/D
-                    tau = min_value(tau, Pe/3)
-                self.test = self.test + tau*dot(velocity, grad(self.test))
-                kwargs['test_function'] = self.test
+    def add_nonconservative_terms(self, *args, **kwargs):
+        self.add_term(HorizontalAdvectionTerm(*args, **kwargs), 'explicit', args[1])
+        self.add_term(HorizontalDiffusionTerm(*args, **kwargs), 'explicit', args[1])
+        self.add_term(SourceTerm(*args, **kwargs), 'source', args[1])
 
-        args = (function_space, depth, options)
-        self.add_terms(*args, **kwargs)
+    def add_conservative_terms(self, *args, **kwargs):
+        self.add_term(ConservativeHorizontalAdvectionTerm(*args, **kwargs), 'explicit', args[1])
+        self.add_term(ConservativeHorizontalDiffusionTerm(*args, **kwargs), 'explicit', args[1])
+        self.add_term(ConservativeSourceTerm(*args, **kwargs), 'source', args[1])
 
-    def add_terms(self, *args, **kwargs):
-        self.add_term(HorizontalAdvectionTerm(*args, **kwargs), 'explicit')
-        self.add_term(HorizontalDiffusionTerm(*args, **kwargs), 'explicit')
-        self.add_term(SourceTerm(*args, **kwargs), 'source')
+    def apply_supg(self, test_function):
+        """
+        Apply SUPG stabilization in the sense of modifying the test function.
+        """
+        unorm = self.options.horizontal_velocity_scale
+        self.cellsize = anisotropic_cell_size(self.function_space.mesh())
+        tau = 0.5*self.cellsize/unorm
+        D = self.options.horizontal_diffusivity_scale
+        Pe = 0.5*unorm*self.cellsize/D
+        tau = conditional(D > 0, min_value(tau, Pe/3), tau)
+        inc = tau*dot(self.velocity, grad(self.test))
+        return test_function + conditional(unorm > 0, inc, 0)
